create cli.rs and make the tests passing
